+++ Gorgias/Gorgias/ProfileReport.cs
@@ -15,7 +15,7 @@ namespace Gorgias
     public partial class ProfileReport
     {
         public int ProfileReportID { get; set; }
-        public int ProfileReportActivityCount { get; set; }
+        public long ProfileReportActivityCount { get; set; }
         public double ProfileReportRevenue { get; set; }
         public int ReportTypeID { get; set; }
         public int ProfileID { get; set; }
